from mc

+++ minicaffe/layers/mnist_generator.h
@@ -24,7 +24,7 @@ private:
     std::ifstream labelFS;
     void loadToMemory();
     std::vector<std::vector<float> > _images;
-    std::vector<float> _labels;
+    std::vector<int> _labels;
     inline int to_int(char* buffer){
         return ((buffer[0] & 0xff) << 24) | ((buffer[1] & 0xff) << 16) |
                ((buffer[2] & 0xff) << 8) | ((buffer[3] & 0xff) << 0);
